requirements.txt
pytest==7.4.0
allure-pytest==2.10.0
requests==2.28.0
